+++ include/open62541pp/detail/traits.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <tuple>
 #include <type_traits>
 
 namespace opcua::detail {
@@ -10,12 +9,6 @@ struct AlwaysFalse : std::false_type {};
 
 template <typename T, typename... Ts>
 using IsOneOf = std::disjunction<std::is_same<T, Ts>...>;
-
-template <typename, typename>
-struct TupleHolds : std::false_type {};
-
-template <typename... Ts, typename T>
-struct TupleHolds<std::tuple<Ts...>, T> : std::bool_constant<(std::is_same_v<Ts, T> || ...)> {};
 
 /// Remove pointer, references and all qualifiers.
 template <typename T>