+++ src/types/DataValue.cpp
@@ -1,31 +1,6 @@
 #include "open62541pp/types/DataValue.h"
 
-#include "../open62541_impl.h"
-
 namespace opcua {
-
-static UA_DataValue fromOptionals(
-    const std::optional<DateTime>& sourceTimestamp,
-    const std::optional<DateTime>& serverTimestamp,
-    const std::optional<uint16_t>& sourcePicoseconds,
-    const std::optional<uint16_t>& serverPicoseconds,
-    const std::optional<StatusCode>& statusCode
-) {
-    return {
-        {},
-        sourceTimestamp.value_or(DateTime{}),
-        serverTimestamp.value_or(DateTime{}),
-        sourcePicoseconds.value_or(uint16_t{}),
-        serverPicoseconds.value_or(uint16_t{}),
-        statusCode.value_or(StatusCode{}),
-        false,
-        sourceTimestamp.has_value(),
-        serverTimestamp.has_value(),
-        sourcePicoseconds.has_value(),
-        serverPicoseconds.has_value(),
-        statusCode.has_value(),
-    };
-}
 
 DataValue::DataValue(
     Variant value,
@@ -34,10 +9,21 @@ DataValue::DataValue(
     std::optional<uint16_t> sourcePicoseconds,
     std::optional<uint16_t> serverPicoseconds,
     std::optional<StatusCode> statusCode
-)
-    : DataValue(fromOptionals(
-          sourceTimestamp, serverTimestamp, sourcePicoseconds, serverPicoseconds, statusCode
-      )) {
+) noexcept
+    : DataValue(UA_DataValue{
+          UA_Variant{},
+          sourceTimestamp.value_or(UA_DateTime{}),
+          serverTimestamp.value_or(UA_DateTime{}),
+          sourcePicoseconds.value_or(uint16_t{}),
+          serverPicoseconds.value_or(uint16_t{}),
+          statusCode.value_or(UA_StatusCode{}),
+          false,
+          sourceTimestamp.has_value(),
+          serverTimestamp.has_value(),
+          sourcePicoseconds.has_value(),
+          serverPicoseconds.has_value(),
+          statusCode.has_value(),
+      }) {
     setValue(std::move(value));
 }
 
@@ -94,36 +80,36 @@ StatusCode DataValue::getStatusCode() const noexcept {
 }
 
 void DataValue::setValue(const Variant& value) {
-    UA_Variant_copy(value.handle(), &handle()->value);
+    asWrapper<Variant>(handle()->value) = value;
     handle()->hasValue = true;
 }
 
-void DataValue::setValue(Variant&& value) {
-    value.swap(handle()->value);
+void DataValue::setValue(Variant&& value) noexcept {
+    asWrapper<Variant>(handle()->value) = std::move(value);
     handle()->hasValue = true;
 }
 
-void DataValue::setSourceTimestamp(DateTime sourceTimestamp) {
+void DataValue::setSourceTimestamp(DateTime sourceTimestamp) noexcept {
     handle()->sourceTimestamp = *sourceTimestamp.handle();
     handle()->hasSourceTimestamp = true;
 }
 
-void DataValue::setServerTimestamp(DateTime serverTimestamp) {
+void DataValue::setServerTimestamp(DateTime serverTimestamp) noexcept {
     handle()->serverTimestamp = *serverTimestamp.handle();
     handle()->hasServerTimestamp = true;
 }
 
-void DataValue::setSourcePicoseconds(uint16_t sourcePicoseconds) {
+void DataValue::setSourcePicoseconds(uint16_t sourcePicoseconds) noexcept {
     handle()->sourcePicoseconds = sourcePicoseconds;
     handle()->hasSourcePicoseconds = true;
 }
 
-void DataValue::setServerPicoseconds(uint16_t serverPicoseconds) {
+void DataValue::setServerPicoseconds(uint16_t serverPicoseconds) noexcept {
     handle()->serverPicoseconds = serverPicoseconds;
     handle()->hasServerPicoseconds = true;
 }
 
-void DataValue::setStatusCode(StatusCode statusCode) {
+void DataValue::setStatusCode(StatusCode statusCode) noexcept {
     handle()->status = statusCode;
     handle()->hasStatus = true;
 }